(feat) O3-3260 Add ability to register feature flags via routes.json
