+++ packages/apps/esm-implementer-tools-app/tsconfig.json
@@ -11,7 +11,17 @@
     "declaration": true,
     "declarationDir": "dist",
     "emitDeclarationOnly": true,
-    "lib": ["dom", "es5", "scripthost", "es2015", "es2015.promise", "es2016.array.include", "es2018", "esnext"]
+    "lib": [
+      "dom",
+      "es5",
+      "scripthost",
+      "es2015",
+      "es2015.promise",
+      "es2016.array.include",
+      "es2018",
+      "es2020",
+      "es2022"
+    ]
   },
   "include": ["src/**/*"]
 }
