+++ packages/apps/esm-offline-tools-app/tsconfig.json
@@ -16,7 +16,8 @@
       "es2015.promise",
       "es2016.array.include",
       "es2018",
-      "es2020"
+      "es2020",
+      "es2022"
     ],
     "noEmit": true
   },
