+++ packages/apps/esm-primary-navigation-app/tsconfig.json
@@ -15,7 +15,8 @@
       "es2015.promise",
       "es2016.array.include",
       "es2018",
-      "es2020"
+      "es2020",
+      "es2022"
     ],
     "noEmit": true
   },
