+++ packages/framework/esm-routes/src/loaders/index.ts
@@ -1,0 +1,3 @@
+export { initializeApp } from './app';
+export { registerApp, finishRegisteringAllApps } from './pages';
+export { tryRegisterExtension } from './components';
