+++ packages/framework/esm-routes/src/public.ts
@@ -1,3 +1,3 @@
 export * from './constants';
-export * from './loaders';
 export * from './routes';
+export { initializeApp } from './loaders';
