+++ packages/framework/esm-styleguide/src/workspaces/action-menu-button/action-menu-button.component.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import classNames from 'classnames';
 import { Button, IconButton } from '@carbon/react';
 import { useLayoutType } from '@openmrs/esm-react-utils';
-import { useWorkspaces } from '../workspaces';
 import styles from './action-menu-button.module.scss';
+import { useWorkspaces } from '../workspaces';
 
 interface TagsProps {
   getIcon: (props: object) => JSX.Element;
