+++ packages/framework/esm-styleguide/src/workspaces/index.ts
@@ -2,5 +2,4 @@ export * from './action-menu-button/action-menu-button.component';
 export * from './container/action-menu.component';
 export * from './container/workspace-container.component';
 export * from './notification/workspace-notification.component';
-export * from './types';
 export * from './workspaces';
