+++ packages/framework/esm-styleguide/src/workspaces/public.ts
@@ -1,5 +1,10 @@
 export * from './action-menu-button/action-menu-button.component';
 export * from './container/workspace-container.component';
-export { type DefaultWorkspaceProps, type CloseWorkspaceOptions } from './types';
 export { closeWorkspace, launchWorkspace, navigateAndLaunchWorkspace, useWorkspaces } from './workspaces';
-export { type OpenWorkspace, type WorkspaceRegistration, type WorkspacesInfo, type Prompt } from './workspaces';
+export {
+  type DefaultWorkspaceProps,
+  type CloseWorkspaceOptions,
+  type OpenWorkspace,
+  type WorkspacesInfo,
+  type Prompt,
+} from './workspaces';
