+++ packages/shell/esm-app-shell/package.json
@@ -52,7 +52,6 @@
     "single-spa": "^6.0.1",
     "swc-loader": "^0.2.3",
     "swr": "^2.2.2",
-    "systemjs": "^6.8.3",
     "webpack": "^5.88.0",
     "webpack-pwa-manifest": "^4.3.0",
     "workbox-core": "^6.1.5",
